scaled integer instead of float for servo and pot config. uint16 for id

+++ Orion_FW/System/Core/Inc/System.h
@@ -47,7 +47,7 @@ public:
 	static void handle_meissa_output_cmd(uint8_t sender_id, LaserPacket* packet);
 
 private:
-	static uint32_t node_id;
+	static uint16_t node_id;
 };
